WIP: Shifted rotates into modules
Some of the components are built in use-orientation and some are built
in print-orientation.  The mix is a hodge-podge.  Mod the modules to all
be print-oriented for consistency.

This seems like it would be better done if all the components were drawn
in print-position rather than rotated there.  This would require some
shifting around of the coordinates, but then it should be cleaner. Leave
this as a TODO for the reader.